refactor: setting ui
